doing the monte carlo tree search algorithm thing

+++ src/main/java/github/thewinner958/game/GameSetup.java
@@ -7,22 +7,22 @@ public class GameSetup {
     private final String charX;
     private final String charO;
     private final String charEmpty;
-    private final Long columns;
-    private final Long rows;
+    private final Integer columns;
+    private final Integer rows;
 
     public GameSetup() {
         charX = "X";
         charO = "O";
         charEmpty = "_";
-        columns = 3L;
-        rows = 3L;
+        columns = 3;
+        rows = 3;
     }
 
-    public GameSetup(String charX, String charO, String charEmpty, Long columns, Long rows) {
+    public GameSetup(String charX, String charO, String charEmpty, Integer columns, Integer rows) {
         this.charX = charX;
         this.charO = charO;
         this.charEmpty = charEmpty;
-        if (columns < 3L || rows < 3L) throw new RuntimeException("You can't have a board less than 3x3");
+        if (columns < 3 || rows < 3) throw new RuntimeException("You can't have a board less than 3x3");
         this.columns = columns;
         this.rows = rows;
     }
